bug 80947
git-svn-id: http://svn.cern.ch/guest/lhcb/Phys/trunk@122481 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 018352c0b603cf05195fc454c29a9b59f7c909ef

+++ Phys/Phys/DaVinciTools/src/DaVinciInit.cpp
@@ -91,7 +91,7 @@ StatusCode DaVinciInit::execute() {
   
   counter("Events")++ ;
   unsigned int nev = counter("Events").nEntries() ;
-  int mem = System::virtualMemory();
+  unsigned long mem = System::virtualMemory();
 
   if (msgLevel(MSG::DEBUG)) debug() << nev << " memory: " << mem << " KB" << endmsg ;
